+++ CHANGELOG.md
@@ -4,6 +4,7 @@ master
 # 4.2.2
 
 * Require i18n ~> 0.8.0 to handle 0.7.0 security issue.
+* Allow Bundler 2
 
 # 4.2.1