+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ===
 
+* Only SassC from now on.
 * Add `--dry-run` to run a build, but skip outputting to disk.
 * Incremental builds: `--track-dependencies` and `--only-changed` flags (#2220)
 * Remove Rack support in favor of `resource.filters << proc { |oldbody| newbody }`