+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ===
 
+* Update activesupport to 5.x and padrino to 0.14.x
 * Only SassC from now on.
 * Add `--dry-run` to run a build, but skip outputting to disk.
 * Incremental builds: `--track-dependencies` and `--only-changed` flags (#2220)